requirement.txt
discord
discord.py
BeautifulSoup4
requests
python-dotenv